backend/requirements.txt
fastapi
python-dotenv
google-generativeai
deep-translator
uvicorn
langchain
langchain-community
pydantic
PyMuPDF
faiss-cpu
